pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc-taxi-clickhouse-etl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "astroid>=4.1.2",
    "boto3>=1.43.2",
    "minio>=7.2.20",
    "python-dotenv>=1.2.2",
    "requests>=2.31",
    "pytest>=7.0",
    "hypothesis>=6.0",
    "mutmut>=3.0",
    "atheris>=3.0",
    "coverage>=7.0",
    "responses>=0.23",
    "vcrpy>=5.0",
    "testcontainers>=4.5",
    "radon>=6.0",
    "xenon>=0.9",
    "vulture>=2.0",
    "cohesion>=1.0",
]

[tool.hypothesis]
max_examples = 30
deadline = 2000

[tool.setuptools.packages.find]
exclude = ["mutants*", "data*"]

[tool.vulture]
ignore_names = [
    "exc_tb",
    "exc_val",
    "fake_parquet_content",
    "mock_200",
    "mock_404",
    "mock_500",
    "side_effect",
    "dirs",
]

[tool.mutmut]
paths_to_mutate = ["extract", "upload", "etl"]
do_not_mutate = ["*/tests/*", "*/conftest.py", "*/test_*.py", "*/__init__.py"]
runner = "python -m pytest extract/tests/downloaders/ upload/tests/ etl/tests/ -q --ignore=extract/tests/downloaders/test_fuzz.py --ignore=extract/tests/downloaders/test_e2e.py --ignore=extract/tests/downloaders/test_e2e_errors.py --ignore=extract/tests/downloaders/test_e2e_retry.py -k \"not e2e_client\""
exclude_dirs = ["__pycache__", ".venv", "mutants"]

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
